MysteryiousHounslow/ml_pipeline/requirements.txt
psycopg2-binary>=2.9.6
pgvector>=0.2.3
xgboost>=1.7.0
ortools>=9.6.2534
pygad>=3.2.0
gql>=3.0.0a5
requests>=2.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0